% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_the_line.R
\name{mark}
\alias{mark}
\title{Mark Active Line}
\usage{
mark()
}
\description{
Marks the active line in the RStudio Source Editor.
}
\details{
The RStudio IDE has two keyboard shortcuts associated with the
\code{Ctrl+Shift+L} keyboard binding: one marks the active line in the source
editor, and the other launches \code{devtools::load_all()} whenever the \code{.Rproj}
file specifies \code{BuildType: Package}.

During package development, the latter masks the former. Unfortunately,
the RStudio IDE does not expose the line marking shortcut, making it
impossible to assign a new keyboard binding to it.

This function mimics the behaviour of the line marking shortcut. The package
then exposes this functionality to the RStudio IDE via an addin that can be
used to set the new keyboard shortcut.
}
